## Linked issue (required)
<!-- Use a closing keyword so it shows up in Development -->
Fixes #49

## What changed
- sun skndirectly

## Why
-  ease of use

## How to test
- run it

+++ inc/erf.h
@@ -14,4 +14,4 @@ struct stats
     double mean;
 };
 
-struct stats fit_erf(uint64_t* samples, size_t size);
+struct stats fit_erf(double* samples, size_t size);
